Make now stable variable instead of function (#6)
CEL expects expressions to be pure, which the `now()` function we
previously introduced would violate. To avoid this, and to align with
other implementations, we're going to make now a global variable
instead, so that it may be consistent within a single expression. The
value can be lazily resolved in an activation to avoid overhead of
resolving this variable on every expression.

Updated conformance tests to ensure `now == now` within an expression.