Add conformance tests for ignore_empty (#126)
Following #124, this adds conformance tests that ensure the
`ignore_empty` rules is consistent with `required`. Mainly, if a field
cannot differentiate between unset and the zero value (i.e., is not
nullable), then this rule applies; effectively this is just repeated and
map fields, as well as non-optional proto3 scalar outside of a oneof.

Running against `protovalidate-go`, a few places where `ignore_empty`
should be a noop ends up disabling evaluations. This will be a minor
follow-up fix in that library. Patches for the other libraries will
follow to bring them into conformance with these tests and those in
#124.