Deprecate skipped rule + add IGNORE_ALWAYS to ignore rule + more explicit naming (#172)
After some internal discussions, the patch contains the following
changes around the `ignore` field constraint:

- Only two "defined" words in the documentation around ignore/required
fields are "populated" (would be present in the wire format) and
"nullable" (what the proto docs call "explicit presence"). This is so
that we don't confuse these already nuanced definitions with the equally
nuanced proto field presence.

- Deprecated the `skipped` field constraint, replacing it with
`IGNORE_ALWAYS` as these behaviors were already mutually exclusive.
`required` will not be rolled in because that is an actual rule (that
checks if a field is populated) and not a modifier of the validator's
behavior.

- rename `IGNORE_EMPTY` and `IGNORE_DEFAULT` to `IGNORE_IF_UNPOPULATED`
and `IGNORE_IF_DEFAULT_VALUE`, respectively, to improve the
self-documenting behavior of these already confusing names.