Optimize Bazel CI performance (#312)
This should improve performance for the Bazel CI workflow, and ensure
that it is cached more often. We will need to manually clear caches
periodically though, since the approach used to save/restore the bazel
cache is naive. (This is also true for the main CI rule, unfortunately.
Maybe there is a more advanced Bazel cache integration for GitHub
Actions.)