Move CEL extension documentation to buf.build/docs (#318)
*Note: If approved, please do not merge. The buf.build/docs page isn't
live yet. I'll coordinate the order of operations once all PRs are
approved.*

This PR removes the content within the CEL overloads section of
`cel.md`, replacing it with a notice that the documentation has moved to
buf.build/docs and a link to the new location. It also updates the
internal link within this page that linked to the anchor for this
section to go directly to the new location.