Add additional tests for map keys (#349)
This adds additional conformance tests for map keys with various scalar
types. Previously, we only tested `int32` as a scalar type for a map
key. This exposed a
[bug](https://github.com/bufbuild/protovalidate-go/pull/219) with
int32/int64 types in Go.

In the interest of completeness, this expands on the scalar type key by
adding tests for:
* `int32`
* `int64`
* `uint32`
* `uint64`
* `bool`
* `string`

Note this doesn't add tests for `fixed32/64` or `sfixed32/64` since
those should be handled by the `uint32/uint64` and `int32/int64` tests
respectively.