Rename constraints to rules (#354)
Right now we co-mingle the terms "Rule" and "Constraint", and to a first
approximation it is a distinction without a difference. It will be
better for users to standardize on using a single term, both in docs and
in code.

This PR proposes standardizing on "Rule" because:

1. This would be a breaking change in either case, but choosing "Rule"
minimizes the impact.
2. Rule is a simple/short word, and
https://github.com/bufbuild/protovalidate/pull/317 (so this is a natural
extension of that).

**This is a breaking change.**

Replaces https://github.com/bufbuild/protovalidate/pull/338