Clarify requried behavior for map keys/values and repeated items (#360)
Internal report that there's some confusion about the behavior of when
`FieldRules.map.keys.required`, `FieldRules.map.values.required`, and
`FieldRules.repeated.items.required` are set to `true`. For maps, the KV
pairs are explicitly set; while technically the key or value could be
elided from the wire format to keep the size down, the field is always
considered "set." Likewise, repeated items are always encoded, even if
they are the zero value; it is impossible to have an unset repeated item
(including messages).

This patch clarifies this in the `required` docs and adds a few
conformance tests to enforce it. I kept it only to the proto3 required
cases. I figured it would be overkill to take this further. Tested on
protovalidate-go main (as of writing this).