Upgrade to use v2 for buf.yaml (#366)
This is a small cleanup PR to convert the repo to use the v2 config for
`buf.yaml`. It deletes the individual files in the proto directory in
favor of a single one at the root which defines both modules.

Also deleted `buf.work.yaml` and `buf.lock` since deps for local modules
are handled by Buf.

Note: should be merged _after_
https://github.com/bufbuild/modules/pull/694