Add new `MessageOneof` rule (#377)
This adds new message rule `oneof` which allows for oneof semantics
without the ceremony of Protobuf `oneof`.

Consumers can specify a list of field names in a message that constitute
the `oneof` and can also specify an optional `required` flag, which
enforces that exactly one of the fields _must_ be set. If `required` is
false, it enforces that zero-or-one of specified fields _can_ be set.