Rename `IGNORE_IF_UNPOPULATED` to `IGNORE_IF_ZERO_VALUE` (#397)
`IGNORE_IF_UNPOPULATED` is a no-op for fields that track presence. It
behaves the same as the default for them, and is only useful for fields
that don't track presence. It reads "unpopulated" for fields that don't
have a logical unpopulated state, yet it is only relevant for them. This
is confusing.

To address this we are renaming the field to `IGNORE_IF_ZERO_VALUE`
along with a note to indicate it is a no-op for fields that track
presence.

While this is a breaking change, there is no change in the behavior.
Migrating is as simple as updating all occurrences of
`IGNORE_IF_UNPOPULATED` to `IGNORE_IF_ZERO_VALUE` and everything will
continue to work as it was working before.

Signed-off-by: Sri Krishna <skrishna@buf.build>