Update to new BCR workflow (#402)
The GitHub App for BCR publishing we implemented in Q1 is now deprecated
in favor of a new approach using GitHub workflows. I think the main
reason for this is to support SLSA attestation for the BCR. Anyway, a
few changes needed to be made:

- Added myself as a maintainer; not _strictly_ necessary, but it's
useful.
- Remove "Create Release Draft" workflow; we have to use the BCR release
workflow in order to pass attestation.
- Add "Publish to BCR" action on release. This does the work that the
publish to BCR app did, and creates a PR on the BCR repo when a release
is published.
- Rename the "Create Release Tag" workflow to just "Release". It still
just creates a draft, but the workflow names were confusing before.
- Update the Release workflow to use the BCR release workflow. Add
necessary permissions for attestation.
- Update release_prep.sh to output the release notes snippet to
`stdout`. Ensure all other commands that might output to `stdout` are
redirected to `stderr`.
- Update the `RELEASING.md` documentation.

We will need to test this by making a new minor release.