Add tests for `required` on implicit presence fields (#438)
This PR clarifies behavior for the `required` rule on fields with
implicit presence: If the rule is violated, other rules on the field are
not evaluated.

For a concrete example:

```proto
syntax="proto3";
message Message {
  string hello = 1 [
    (buf.validate.field).required = true,
    (buf.validate.field).string.min_len = 1
  ];
}
```

If the field `hello` has a zero value, the validator should return a
violation for the `required` rule, but not for the `string.min_len`
rule.

This PR adds 9 tests to verify the behavior. Also see
https://github.com/bufbuild/protovalidate-es/issues/107.