Pin tool versions and setup bazelisk (#445)
[bazelisk](https://github.com/bazelbuild/bazelisk) is a widely used
version management tool for Bazel. Most Bazel projects expect users to
use it. Protovalidate on the other hand is not primarily a Bazel repo,
the targets are added to be used in
[protovalidate-cc](https://github.com/bufbuild/protovalidate-cc) and
other Bazel projects. Setting up the bazelisk tool will be make it
easier for users run lint and tests.

Closes #431