Add `cel_expression` (#432)
Whenever I define a CEL rule, I am constantly annoyed that I have to
define an `id` and a `message` when the expression contains all the
context I need. I want to be able to do this:

```protobuf
message Invoice {
  option (buf.validate.message).cel = {expression: "this.sub_total.currency_code == this.taxes.currency_code"};

  money.v1.Money sub_total = 5 [(buf.validate.field).required = true];
  money.v1.Money taxes = 6 [(buf.validate.field).required = true];
}
```

Instead, I'm all but forced to do this:

```protobuf
message Invoice {
  option (buf.validate.message).cel = {
    id: "same_currency_code"
    message: "sub_total and taxes cannot have different currency codes"
    expression: "this.sub_total.currency_code == this.taxes.currency_code"
  };

  money.v1.Money sub_total = 5 [(buf.validate.field).required = true];
  money.v1.Money taxes = 6 [(buf.validate.field).required = true];
}
```

I don't understand why I have to come up with a unique ID, and why I
have to pay the mental overhead of coming up with a human-readable
message when all the context I need is right in the expression.

It's also super annoying to me that `(buf.validate.field).cel` isn't
CEL. It's a message that then has a field `expression` that is CEL. All
I actually want to do is this:

```protobuf
message Invoice {
  option (buf.validate.message).cel = "this.sub_total.currency_code == this.taxes.currency_code";
}
```

As a next-best, I'd want to do this:

```protobuf
message Invoice {
  option (buf.validate.message).cel.expression = "this.sub_total.currency_code == this.taxes.currency_code";
}
```

But this is a Protobuf compile error because
`(buf.validate.message).cel` is a repeated message.

In hindsight, I wish we did:

```protobuf
message MessageRules {
  repeated string cel = 1;
  repeated Rule rule = 2;
}
```

And then I could do the full version with `(buf.validate.message).rule`
and the simple version with `(buf.validate.message).cel`, which also
matches what I'd expect (`rule` is a `Rule`, `cel` is a CEL expression).
But we can't do that anymore.

Perhaps we could do what this PR proposes. Then I could do:

```protobuf
message Invoice {
  option (buf.validate.message).cel_expression = "this.sub_total.currency_code == this.taxes.currency_code";
}
```

This would need to be done for fields, predefined, etc, and new
conformance tests added.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: Christoph Hoopmann <choopm@0pointer.org>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Spencer Connaughton <spencer@spencerconnaughton.com>
Co-authored-by: Christoph Hoopmann <christophhoopmann@gmail.com>
Co-authored-by: Stefan VanBuren <svanburen@buf.build>
Co-authored-by: Timo Stamm <ts@timostamm.de>
Co-authored-by: Chris Roche <github@rodaine.com>
Co-authored-by: Mimu <52129983+fchimpan@users.noreply.github.com>
Co-authored-by: Sri Krishna <93153132+srikrsna-buf@users.noreply.github.com>
Co-authored-by: Sri Krishna <skrishna@buf.build>